+++ src/agent/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * Update `StreamListener` to use `DeltaEvent` and `TextDelta` instead of `ChunkEvent` and raw strings
+ * Update `StreamListener` to react to `ToolCallComplete` instead of `ToolCallResult`
 
 0.4
 ---